Abstracted PaymentViewModel to its own function and reset CaseService logic back to original logic
