+++ api/src/middlewares/services/CaseService/types/CaseService.ts
@@ -1,13 +1,15 @@
 import { AlertJob } from "../../utils/types";
 import { QueueService } from "../../QueueService";
 import { FormField } from "../../../../types/FormField";
-import { FormType, PayMetadata } from "../../../../types/FormDataBody";
+import { FormType } from "../../../../types/FormDataBody";
 import { SESSendJob } from "./SESSendJob";
 import { ProcessQueueData } from "./ProcessQueueData";
-import { PaymentViewModel } from "./PaymentViewModel";
+import { PaymentData } from "./PaymentData";
 import { ProcessQueueDataInput } from "./ProcessQueueDataInput";
+import { Logger } from "pino";
 
 export interface CaseService {
+  logger: Logger;
   queueService: QueueService;
 
   sendToProcessQueue(data: ProcessQueueData): Promise<string>;
@@ -17,18 +19,12 @@ export interface CaseService {
   /**
    * Builds the email body
    */
-  getEmailBody(data: { fields: FormField[]; payment?: PaymentViewModel; reference: string }, type: FormType): string;
+  getEmailBody(data: { fields: FormField[]; payment?: PaymentData; reference: string }, type: FormType): string;
 
   /**
    * builds the data required for the SES_PROCESS job
    */
   buildProcessQueueData(input: ProcessQueueDataInput): ProcessQueueData;
-
-  /**
-   * Parses the payment data in metadata
-   * TODO:- refactor into own class or function
-   */
-  paymentViewModel(payment: PayMetadata | undefined, country: string): PaymentViewModel | undefined;
 
   /**
    * builds the data required for the SES_SEND job
