+++ api/src/middlewares/services/CaseService/types/index.ts
@@ -1,5 +1,5 @@
 export { FormMetadata, CertifyCopyFormMetadata, MarriageFormMetadata } from "./FormMetadata";
-export { PaymentViewModel } from "./PaymentViewModel";
+export { PaymentData } from "./PaymentData";
 export { ProcessQueueData, CertifyCopyProcessQueueData, MarriageProcessQueueData } from "./ProcessQueueData";
 export { SESSendJob } from "./SESSendJob";
 export { CaseService } from "./CaseService";
